resin-filesystem-expand: Remove udev settle dependency
We introduced this dependency in order to make sure the filesystem labels are
in place when running the mounts. We were forced to not use a mount unit due to
an issue when running these units in a container. See
06befb503ba25a244027861b367e47fa2fb43630 for more info.

This though was proven to break the system when settle fails due to some
devices timing out at boot. This was seen on USB modem for example. And in any
way this was an overkill as we are not interested for all the devices to settle
but only the labels to be in place. We simplify this by just waiting for the
label to "appear" for 10 seconds which *should be* enough for now.  This is a
workaround for now and we intend to get back to mounts as soon as we find a
solution to the issue mentioned above.

Signed-off-by: Andrei Gherzan <andrei@resin.io>